cleaned up the branch somewhat
